smart-receipt-tracker/api/requirements.txt
azure-functions
azure-ai-formrecognizer==3.3.0
